Added LED download indicator and restructured download test for TLS mode
