+++ TESTS/COMMON/download_test.h
@@ -1,0 +1,5 @@
+
+
+static void socket_event(void);
+
+void download_test(size_t size, NetworkInterface* interface, bool tls_mode);
